+++ hpproj/__init__.py
@@ -20,6 +20,7 @@ HealPix Projection
 
 __version__ = '0.7.4'
 
+from .wcs_helper import *
 from .hp_helper import *
 from .cutsky import *
 from .visu import *